Added login functionality
- login button on index
- logout option on studenlist
- edit/delete/new doesn't show up unless logged in

+++ students.php
@@ -1,3 +1,5 @@
+<?php session_start(); ?>
+
 <?php
 $customCSS = "<link rel='stylesheet' href='css/styles.css'>";
 $useNav = true;
@@ -13,6 +15,7 @@ $result = mysqli_query($connection, "SELECT * FROM studentlist ORDER BY fullName
     <div class="container">
 
         <h1>Students</h1>
+
         <article>
 
             <?php
@@ -55,8 +58,10 @@ $result = mysqli_query($connection, "SELECT * FROM studentlist ORDER BY fullName
                     </div>
                     <div class="link">
                         <div class="edit-delete">
-                            <a href="edit.php?id=<?php echo $row['id'];?>">Edit</a>
-                            <a onclick="return confirm('Are you sure you want to delete: <?php echo $row["fullName"];?>?')" href="delete.php?id=<?php echo $row['id']; ?>">Delete</a>
+                            <?php if(isset($_SESSION['username'])) { ?>
+                                <a href="edit.php?id=<?php echo $row['id'];?>">Edit</a>
+                                <a onclick="return confirm('Are you sure you want to delete: <?php echo $row["fullName"];?>?')" href="delete.php?id=<?php echo $row['id']; ?>">Delete</a>
+                            <?php } ?>    
                         </div>
                         <div class="view-page">
                             <a href="<?php echo $row['link'];?>" class="button-link">View <?php echo $row['fullName'];?>'s Page</a>
